% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active.extension.R
\name{active.extension}
\alias{active.extension}
\title{Enable active extension portfolios}
\usage{
active.extension(model, up = 130, down = 30)
}
\arguments{
\item{model}{the portfolio.model to activate}

\item{up}{percentage long (e.g. 130)}

\item{down}{percentage short (e.g. 30)}
}
\value{
portfolio.model with active extension enabled
}
\description{
\code{active.extension} adds corresponding long/short constraints for a 
diverse set of active extension portfolios (e.g. 130/30 portfolios)
}
\author{
Ronald Hochreiter, \email{ronald@algorithmic.finance}
}
