% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_portfolio.default.R
\name{aux_portfolio.default}
\alias{aux_portfolio.default}
\title{Set portfolio.model default values}
\usage{
aux_portfolio.default(model)
}
\arguments{
\item{model}{the portfolio.model to be reset}
}
\value{
a portfolio.model with all default values set
}
\description{
\code{aux_portfolio.default} sets portfolio.model default values
}
\author{
Ronald Hochreiter, \email{ronald@algorithmic.finance}
}
