% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_risk.alias.R
\name{aux_risk.alias}
\alias{aux_risk.alias}
\title{Convert risk alias names to internal names}
\usage{
aux_risk.alias(risk)
}
\arguments{
\item{risk}{the risk name to be standardized}
}
\value{
the standardized risk name (if any)
}
\description{
\code{aux_risk.alias} converts risk alias names to internal names
}
\author{
Ronald Hochreiter, \email{ronald@algorithmic.finance}
}
