% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_simulate.scenarios.R
\name{aux_simulate.scenarios}
\alias{aux_simulate.scenarios}
\title{Simulate a multivariate-normal scenario.set}
\usage{
aux_simulate.scenarios(mu, Sigma, n = 1000, seed = 280277)
}
\arguments{
\item{mu}{mean vector of asset returns}

\item{Sigma}{covariance matrix of asset returns}

\item{n}{number of scenarios to simulate (default 1000)}

\item{seed}{random number seed (default 280277)}
}
\value{
A scenario set `simulation` with mean `mu` and covariance `Sigma`
}
\description{
\code{aux_simulate.scenarios} simulates a scenario.set given a mean vector and
a covariance matrix using mvrnorm of the MASS package
}
\author{
Ronald Hochreiter, \email{ronald@algorithmic.finance}
}
