% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.constraint.eq.R
\name{linear.constraint.eq}
\alias{linear.constraint.eq}
\title{Create or update a vector-based linear equality constraint set}
\usage{
linear.constraint.eq(constraints.linear, range, beq, factors = NULL)
}
\arguments{
\item{constraints.linear}{the current set of equality constraints}

\item{range}{the range of the variables to set (default 1 if factors is NULL)}

\item{beq}{right-hand side scalar}

\item{factors}{values to set for each variable in the given range}
}
\value{
the new (updated) set of equality constraints
}
\description{
\code{linear.constraint.eq} creates a vector-based linear equality
constraint: Aeq(range) * factors == beq
}
\author{
Ronald Hochreiter, \email{ronald@algorithmic.finance}
}
