% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long.only.R
\name{long.only}
\alias{long.only}
\title{Disable active extension portfolios}
\usage{
long.only(model)
}
\arguments{
\item{model}{the portfolio.model to deactivate active extensions}
}
\value{
portfolio.model with active extension disabled
}
\description{
\code{long.only} switches a portfolio.model back to long-only by
disabling the active extension
}
\author{
Ronald Hochreiter, \email{ronald@algorithmic.finance}
}
