% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum.R
\name{momentum}
\alias{momentum}
\title{Set momentum parameters for a portfolio.model}
\usage{
momentum(model, n_momentum, n_momentum.short = NULL)
}
\arguments{
\item{model}{the portfolio.model to be changed}

\item{n_momentum}{amount of momentum assets long}

\item{n_momentum.short}{amount of momentum assets short}
}
\value{
the adapted portfolio.model
}
\description{
\code{momentum} sets a new alpha for VaR and Expected Shortfall
}
\author{
Ronald Hochreiter, \email{ronald@algorithmic.finance}
}
