% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{objective}
\alias{objective}
\title{Set new objective of a portfolio.model}
\usage{
objective(model, objective = "markowitz")
}
\arguments{
\item{model}{the portfolio.model to be changed}

\item{objective}{the new objective}
}
\value{
the adapted portfolio.model
}
\description{
\code{objective} sets a new objective for VaR and Expected Shortfall
}
\examples{
data(sp100w17av30s)
model <- portfolio.model(scenario.set)
mad <- optimal.portfolio(objective(model, "mad"))

}
\author{
Ronald Hochreiter, \email{ronald@algorithmic.finance}
}
