% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal.portfolio.1overN.R
\name{optimal.portfolio.1overN}
\alias{optimal.portfolio.1overN}
\title{1 over N portfolio}
\usage{
optimal.portfolio.1overN(model)
}
\arguments{
\item{model}{the portfolio.model to compute the portfolio of}
}
\value{
the portfolio.model including the newly computed optimal portfolio
}
\description{
\code{optimal.portfolio.1overN} adds a 1 over N portfolio to the portfolio.model
}
\author{
Ronald Hochreiter, \email{ronald@algorithmic.finance}
}
