% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/optimal.portfolio.expected.shortfall.long.short.R
\name{optimal.portfolio.expected.shortfall.long.short}
\alias{optimal.portfolio.expected.shortfall.long.short}
\title{Portfolio Optimization minimizing Conditional Value at Risk (CVaR)
with active extensions}
\usage{
optimal.portfolio.expected.shortfall.long.short(model)
}
\arguments{
\item{model}{the portfolio.model to compute the portfolio of}
}
\value{
the portfolio.model including the newly computed optimal portfolio
}
\description{
\code{optimal.portfolio.expected.shortfall.long.short} conducts a Portfolio 
Optimization minimizing Conditional Value at Risk (CVaR) based on 
Rockafellar and Uryasev (2001) with active extensions
}
\author{
Ronald Hochreiter, \email{ronald@algorithmic.finance}
}
