% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal.portfolio.expected.shortfall.R
\name{optimal.portfolio.expected.shortfall}
\alias{optimal.portfolio.expected.shortfall}
\title{Portfolio Optimization minimizing Conditional Value at Risk (CVaR)}
\usage{
optimal.portfolio.expected.shortfall(model)
}
\arguments{
\item{model}{the portfolio.model to compute the portfolio of}
}
\value{
the portfolio.model including the newly computed optimal portfolio
}
\description{
\code{optimal.portfolio.expected.shortfall} conducts a Portfolio Optimization
minimizing Conditional Value at Risk (CVaR) based on Rockafellar and 
Uryasev (2001)
}
\author{
Ronald Hochreiter, \email{ronald@algorithmic.finance}
}
