% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal.portfolio.mad.long.short.R
\name{optimal.portfolio.mad.long.short}
\alias{optimal.portfolio.mad.long.short}
\title{Portfolio Optimization minimizing MAD (Active Extension)}
\usage{
optimal.portfolio.mad.long.short(model)
}
\arguments{
\item{model}{the portfolio.model to compute the portfolio of}
}
\value{
the portfolio.model including the newly computed optimal portfolio
}
\description{
\code{optimal.portfolio.mad.long.short} conducts a Portfolio Optimization minimizing Mean 
Absolute Deviation (MAD) based on Konno and Yamazaki (1991) including an
active extension
}
\author{
Ronald Hochreiter, \email{ronald@algorithmic.finance}
}
