% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal.portfolio.mad.R
\name{optimal.portfolio.mad}
\alias{optimal.portfolio.mad}
\title{Portfolio Optimization minimizing MAD}
\usage{
optimal.portfolio.mad(model)
}
\arguments{
\item{model}{the portfolio.model to compute the portfolio of}
}
\value{
the portfolio.model including the newly computed optimal portfolio
}
\description{
\code{optimal.portfolio.mad} conducts a Portfolio Optimization minimizing Mean 
Absolute Deviation (MAD) based on Konno and Yamazaki (1991)
}
\author{
Ronald Hochreiter, \email{ronald@algorithmic.finance}
}
