% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal.portfolio.markowitz.R
\name{optimal.portfolio.markowitz}
\alias{optimal.portfolio.markowitz}
\title{Portfolio Optimization minimizing Standard Deviation}
\usage{
optimal.portfolio.markowitz(model)
}
\arguments{
\item{model}{the portfolio.model to compute the portfolio of}
}
\value{
the portfolio.model including the newly computed optimal portfolio
}
\description{
\code{portfolio.weights} conducts a Portfolio Optimization minimizing Standard 
Deviation based on Markowitz (1952).
}
\author{
Ronald Hochreiter, \email{ronald@algorithmic.finance}
}
