% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal.portfolio.momentum.R
\name{optimal.portfolio.momentum}
\alias{optimal.portfolio.momentum}
\title{Momentum portfolio including momentum for active extensions}
\usage{
optimal.portfolio.momentum(model)
}
\arguments{
\item{model}{the portfolio.model to compute the portfolio of}
}
\value{
the portfolio.model including the newly computed optimal portfolio
}
\description{
\code{optimal.portfolio.momentum} adds a momentum portfolio to the portfolio.model
}
\author{
Ronald Hochreiter, \email{ronald@algorithmic.finance}
}
