% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal.portfolio.reward.R
\name{optimal.portfolio.reward}
\alias{optimal.portfolio.reward}
\title{Compute maximum/minimum return portfolio given the constraints}
\usage{
optimal.portfolio.reward(model)
}
\arguments{
\item{model}{the portfolio.model to compute the portfolio of}
}
\value{
the portfolio.model including the newly computed optimal portfolio
}
\description{
\code{optimal.portfolio.reward} computes a maximum/minimum return portfolio 
given the constraints
}
\author{
Ronald Hochreiter, \email{ronald@algorithmic.finance}
}
