% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/po.tutorial.R
\name{po.tutorial}
\alias{po.tutorial}
\title{Open a specific portfolio.optimization package tutorial}
\usage{
po.tutorial(tutorial = "")
}
\arguments{
\item{tutorial}{name of the tutorial to open}
}
\value{
Nothing if no tutorial specified, otherwise the path to the tutorial.
}
\description{
\code{po.tutorial} returns the filename of a specific portfolio.optimization 
package tutorial. If no tutorial is given or the tutorial is missspelled, a 
list of available tutorials is printed.
}
\examples{
\dontrun{
file.edit(po.tutorial("101"))
file.edit(po.tutorial("compare"))
}

}
\author{
Ronald Hochreiter, \email{ronald@algorithmic.finance}
}
