% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.loss.R
\name{portfolio.loss}
\alias{portfolio.loss}
\alias{l}
\title{Return the loss distribution of the portfolio.model}
\usage{
portfolio.loss(model)

l(model)
}
\arguments{
\item{model}{the portfolio.model to display}
}
\value{
nothing
}
\description{
\code{portfolio.loss} return the loss distribution of the portfolio.model
}
\author{
Ronald Hochreiter, \email{ronald@algorithmic.finance}
}
