% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.model.R
\name{portfolio.model}
\alias{portfolio.model}
\alias{p.mo}
\title{Create a portfolio.model instance (or fix an existing one)}
\usage{
portfolio.model(input = NULL)

p.mo(input = NULL)
}
\arguments{
\item{input}{model, scenario.set or mean.covariance list}
}
\value{
an S3 object of class portfolio.model
}
\description{
\code{portfolio.model} creates a new S3 portfolio.model instance or
fixes an existing one.
}
\author{
Ronald Hochreiter, \email{ronald@algorithmic.finance}
}
