% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.optimization.R
\docType{package}
\name{portfolio.optimization-package}
\alias{portfolio.optimization}
\alias{portfolio.optimization-package}
\title{Contemporary Portfolio Optimization}
\description{
Simplify your portfolio optimization process by applying a contemporary modeling way to model and solve your portfolio problems. While most approaches and packages are rather complicated this one tries to simplify things and is agnostic regarding risk measures as well as optimization solvers. Some of the methods implemented are described by Konno and Yamazaki (1991) <doi:10.1287/mnsc.37.5.519>, Rockafellar and Uryasev (2001) <doi:10.21314/JOR.2000.038> and Markowitz (1952) <doi:10.1111/j.1540-6261.1952.tb01525.x>.
}
\references{
\url{http://www.finance-r.com/}
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.finance-r.com/}
}

}
\author{
Ronald Hochreiter, \email{ronald@hochreiter.net}
}
