% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.weights.R
\name{portfolio.weights}
\alias{portfolio.weights}
\alias{portfolio}
\alias{w}
\alias{weights}
\alias{x}
\title{Return the portfolio weights of a portfolio.model}
\usage{
portfolio.weights(model)

portfolio(model)

w(model)

weights(model)

x(model)
}
\arguments{
\item{model}{the portfolio.model to return the portfolio weights from}
}
\value{
a vector of portfolio weights or NULL if no weights are available yet.
}
\description{
\code{portfolio.weights} return the portfolio weights of a portfolio.model
}
\examples{
data(sp100w17av30s)
portfolio.weights(optimal.portfolio(scenario.set))

}
\author{
Ronald Hochreiter, \email{ronald@algorithmic.finance}
}
