% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.portfolio.model.R
\name{print.portfolio.model}
\alias{print.portfolio.model}
\title{Overload print() for portfolio.model}
\usage{
\method{print}{portfolio.model}(x, ...)
}
\arguments{
\item{x}{the portfolio.model to display}

\item{...}{further parameters}
}
\value{
nothing
}
\description{
\code{print.portfolio.model} overloads print() and outputs some information
about the portfolio.model object
}
\author{
Ronald Hochreiter, \email{ronald@algorithmic.finance}
}
