% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sp100w17}
\alias{sp100w17}
\title{S&P 100 weekly stock returns 2017}
\format{One xts time series object with 251 rows and 101 columns.}
\usage{
data(sp100w17)
}
\description{
A dataset sp100w17 containing the (crude) weekly returns of (almost) all 
S&P 100 stocks of 2017, daily basis (101 stocks, 251 returns).
}
\details{
Furthermore contains a vector sp100w17av with the average trading volume
of all stocks in 2017 - to be used for a subselection.
}
\keyword{datasets}
