% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sp100w17av}
\alias{sp100w17av}
\title{S&P 100 average trading volume over the whole year 2017}
\format{One named numeric vector of length 101.}
\usage{
data(sp100w17)
}
\description{
A vector sp100w17av with the average trading volume
of all stocks in 2017 - to be used e.g. for a subselection.
}
\keyword{datasets}
