% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sp100w17av30s}
\alias{sp100w17av30s}
\alias{scenario.set}
\title{S&P 100 weekly stock returns 2017 of 30 stocks with the highest average
trading volume over the whole year}
\format{A named matrix scenario.set with 251 rows and 30 columns.}
\usage{
data(sp100w17av30s)
}
\description{
A sceario sp100w17 containing the (crude) weekly returns of (almost) all 
S&P 100 stocks of 2017, daily basis (101 stocks, 251 returns).
}
\keyword{datasets}
