% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper.bound.R
\name{upper.bound}
\alias{upper.bound}
\title{Set upper bounds on assets}
\usage{
upper.bound(model, v1 = NULL, v2 = NULL)
}
\arguments{
\item{model}{the portfolio.model to adapt the upper bounds}

\item{v1}{either one upper bound or lower bounds for all assets}

\item{v2}{if not empty then v1 contains the positions (or names) 
and v2 the bounds}
}
\value{
portfolio.model with new upper bounds
}
\description{
\code{upper.bound} sets lower bounds on assets within a portfolio.model
}
\author{
Ronald Hochreiter, \email{ronald@algorithmic.finance}
}
